Changed how I was saving round data and Implemented a form of Settings
The round number and round time are now both saved into a struct called Round and a list of round is need for the results at the end of the test so the user can view all their round times. Then I wanted to make sure I was keeping to the brief with having only 4 colours and since I implemented the Test to have as many colours as a Designer would like the test to have. It wasn't to difficult to implement.