Cleaned code and added a highlighting feature to show which was your best round
This is the last revision Submission of the Stroop Test project I might work on it in the future. I Cleaned the code and Change the way the text works for all the round times. The way I did have it was 10 set Text Gameobjects in the scene and grabbing each of them. Instead I instantiate a prefab Text Game Object and set it position this makes it easier for when the player changes how many rounds there are. Also the Lowest/Best score is highlighted.